git-notion-sync/requirements.txt
GitPython==3.1.43
PyYAML==6.0.2
notion-client==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
tenacity==8.5.0
